refactor: os_fonts.rs のインラインテストを tests/ に移動し日本語コメントを英語化
self-review 指摘対応:
- src/ 内に #[cfg(test)] mod tests を置くルール違反を解消
  → tests/os_fonts.rs に移動（scan_directory, process_entry を pub に昇格）
- 日本語インラインコメント3箇所を英語に翻訳 (Language Rule 準拠)
- tests/os_fonts.rs に process_entry_accepts_ttc_and_otf テストを追加
